Update 05-shell-introduction.md
edited for clarity. removal of some jargon. consistently formatting of code/variables with backticks.
Got rid of the clear section because it is just extra info. Ditto the preparation magic. It is just extra detail which isn't needed and will go wrong for some. Better to spend the time helping people with the key commands. And explaining working directories!